fix type conversion Pt.2

+++ mlsauce/adaopt/_adaoptc.pyx
@@ -430,7 +430,7 @@ def find_kmin_x(x, n_x, k, cache=False):
     
     return (np.asarray(list(res.values())), 
             np.asarray(list(res.keys()), 
-            dtype=np.integer))                    
+            dtype=np.int64))                    
 
 
 # calculate probs test i
